+++ src/index.js
@@ -22,7 +22,7 @@ import './js/search-form.js';
 //=== === = MAIN'S PART = === ===
 //=== === === === === === === ===
 import './js/card-templete';
-//
+import './js/render-movies';
 //
 //
 //
