+++ src/js/render-movies.js
@@ -1,4 +1,17 @@
-// import refs from './refs';
-// import templeteCard from './card-templete';
+import refs from '../js/refs';
+import templeteCard from './card-templete';
+import MovieService from '../js/movie-service';
 
-// console.log(templeteCard);
+const movieService = new MovieService();
+
+//-------- Рендер популярних фільмів -----------------------------
+movieService
+  .getPopularMovies()
+  .then(data => data.results.map(data => renderCollection(data)));
+
+function renderCollection(data) {
+  const card = templeteCard(data);
+
+  refs.moviesCard.insertAdjacentHTML('beforeend', card);
+}
+//----------------------------------------------------------------
