+++ src/partials/main.html
@@ -2,173 +2,7 @@
   <div class="container">
     <section class="movies">
       <h1 class="visually-hidden">search movies</h1>
-      <ul class="movies-card">
-        <!-- <li class="movies-card__item">
-          <a
-            href="#"
-            class="movies-card__link"
-            data-modal-open
-            data-card-id="тут id фільму"
-          >
-            <div class="movies-card__thumb">
-              <img
-                src="https://image.tmdb.org/t/p/w300//wKiOkZTN9lUUUNZLmtnwubZYONg.jpg"
-                alt=""
-                loading="lazy"
-              />
-            </div>
-            <div class="movies-card__content">
-              <h2 class="movies-card__heading">Minions</h2>
-              <p class="movies-card__text">
-                Drama, Action | 2020
-                <span class="movies-card__rating">9.6</span>
-              </p>
-            </div>
-          </a>
-        </li>
-        <li class="movies-card__item">
-          <a
-            href="#"
-            class="movies-card__link"
-            data-model-open
-            data-card-id="тут id фільму"
-          >
-            <div class="movies-card__thumb">
-              <img
-                src="https://image.tmdb.org/t/p/w300//wKiOkZTN9lUUUNZLmtnwubZYONg.jpg"
-                alt=""
-                loading="lazy"
-              />
-            </div>
-            <div class="movies-card__content">
-              <h2 class="movies-card__heading">Minions</h2>
-              <p class="movies-card__text">Drama, Action | 2020</p>
-            </div>
-          </a>
-        </li>
-        <li class="movies-card__item">
-          <a
-            href="#"
-            class="movies-card__link"
-            data-model-open
-            data-card-id="тут id фільму"
-          >
-            <div class="movies-card__thumb">
-              <img
-                src="https://image.tmdb.org/t/p/w300//wKiOkZTN9lUUUNZLmtnwubZYONg.jpg"
-                alt=""
-                loading="lazy"
-              />
-            </div>
-            <div class="movies-card__content">
-              <h2 class="movies-card__heading">Minions</h2>
-              <p class="movies-card__text">Drama, Action | 2020</p>
-            </div>
-          </a>
-        </li>
-        <li class="movies-card__item">
-          <a
-            href="#"
-            class="movies-card__link"
-            data-model-open
-            data-card-id="тут id фільму"
-          >
-            <div class="movies-card__thumb">
-              <img
-                src="https://image.tmdb.org/t/p/w300//wKiOkZTN9lUUUNZLmtnwubZYONg.jpg"
-                alt=""
-                loading="lazy"
-              />
-            </div>
-            <div class="movies-card__content">
-              <h2 class="movies-card__heading">Minions</h2>
-              <p class="movies-card__text">Drama, Action | 2020</p>
-            </div>
-          </a>
-        </li>
-        <li class="movies-card__item">
-          <a
-            href="#"
-            class="movies-card__link"
-            data-model-open
-            data-card-id="тут id фільму"
-          >
-            <div class="movies-card__thumb">
-              <img
-                src="https://image.tmdb.org/t/p/w300//wKiOkZTN9lUUUNZLmtnwubZYONg.jpg"
-                alt=""
-                loading="lazy"
-              />
-            </div>
-            <div class="movies-card__content">
-              <h2 class="movies-card__heading">Minions</h2>
-              <p class="movies-card__text">Drama, Action | 2020</p>
-            </div>
-          </a>
-        </li>
-        <li class="movies-card__item">
-          <a
-            href="#"
-            class="movies-card__link"
-            data-model-open
-            data-card-id="тут id фільму"
-          >
-            <div class="movies-card__thumb">
-              <img
-                src="https://image.tmdb.org/t/p/w300//wKiOkZTN9lUUUNZLmtnwubZYONg.jpg"
-                alt=""
-                loading="lazy"
-              />
-            </div>
-            <div class="movies-card__content">
-              <h2 class="movies-card__heading">Minions</h2>
-              <p class="movies-card__text">Drama, Action | 2020</p>
-            </div>
-          </a>
-        </li>
-        <li class="movies-card__item">
-          <a
-            href="#"
-            class="movies-card__link"
-            data-model-open
-            data-card-id="тут id фільму"
-          >
-            <div class="movies-card__thumb">
-              <img
-                src="https://image.tmdb.org/t/p/w300//wKiOkZTN9lUUUNZLmtnwubZYONg.jpg"
-                alt=""
-                loading="lazy"
-              />
-            </div>
-            <div class="movies-card__content">
-              <h2 class="movies-card__heading">Minions</h2>
-              <p class="movies-card__text">Drama, Action | 2020</p>
-            </div>
-          </a>
-        </li>
-        <li class="movies-card__item">
-          <a
-            href="#"
-            class="movies-card__link"
-            data-model-open
-            data-card-id="тут id фільму"
-          >
-            <div class="movies-card__thumb">
-              <img
-                src="https://image.tmdb.org/t/p/w300//wKiOkZTN9lUUUNZLmtnwubZYONg.jpg"
-                alt=""
-                loading="lazy"
-              />
-            </div>
-            <div class="movies-card__content">
-              <h2 class="movies-card__heading">
-                Minions,Minions,Minions,Minions,Minions
-              </h2>
-              <p class="movies-card__text">Drama, Action | 2020</p>
-            </div>
-          </a>
-        </li> -->
-      </ul>
+      <ul class="movies-card" data-id="moviesCard"></ul>
     </section>
   </div>
 </main>
